+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### v0.66.4
+- Fix for slight rightward shift of centered point text (Issue #84).
+
+### v0.66.3
+- Fix for bug causing misalignment of columns of text at some browser zooms, affecting Chrome and Safari (Issue #83).
+
 ### v0.66.2
 - Fix: Prevent creation of single-layer image files with duplicate names.
 - Prevent export of single-layer SVG files containing no artwork.
@@ -7,7 +13,6 @@
 
 ### v0.66.0
 - Export content of layers with ":svg" appended to their layer name as SVG files.
-
 
 ### v0.65.6
 - Prevent paths in SVG output from scaling by adding CSS to the .svg file.